Convert from controller to resource
