--- omod/src/main/java/org/openmrs/module/emrapi/web/controller/ProcedureController.java
@@ -1,269 +1,0 @@
-/**
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
- * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
- *
- * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
- * graphic logo is a trademark of OpenMRS Inc.
- */
-package org.openmrs.module.emrapi.web.controller;
-
-import org.apache.commons.lang.StringUtils;
-import org.openmrs.Concept;
-import org.openmrs.Encounter;
-import org.openmrs.Patient;
-import org.openmrs.api.ConceptService;
-import org.openmrs.api.EncounterService;
-import org.openmrs.api.PatientService;
-import org.openmrs.module.emrapi.procedure.Procedure;
-import org.openmrs.module.emrapi.procedure.ProcedureService;
-import org.openmrs.module.emrapi.web.dto.ProcedureDTO;
-import org.openmrs.module.webservices.rest.web.v1_0.controller.BaseRestController;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * REST controller for managing Procedure records.
- * Provides endpoints for creating, retrieving, updating, and voiding procedures.
- */
-@Controller
-@RequestMapping(value = "/rest/**/emrapi/procedure")
-public class ProcedureController extends BaseRestController {
-   
-   @Autowired
-   private ProcedureService procedureService;
-   
-   @Autowired
-   private PatientService patientService;
-   
-   @Autowired
-   private ConceptService conceptService;
-   
-   @Autowired
-   private EncounterService encounterService;
-   
-   /**
-    * Gets all procedures for a patient.
-    *
-    * GET /rest/v1/emrapi/procedure?patient={uuid}
-    *
-    * @param patientUuid the patient UUID
-    * @return list of procedure DTOs
-    */
-   @RequestMapping(method = RequestMethod.GET)
-   @ResponseBody
-   public List<ProcedureDTO> getProcedures(
-           @RequestParam("patient") String patientUuid) {
-      
-      Patient patient = patientService.getPatientByUuid(patientUuid);
-      if (patient == null) {
-         throw new IllegalArgumentException("Patient not found with UUID: " + patientUuid);
-      }
-      
-      List<Procedure> procedures = procedureService.getProceduresByPatient(patient);
-      
-      List<ProcedureDTO> dtos = new ArrayList<>();
-      for (Procedure procedure : procedures) {
-         dtos.add(toDTO(procedure));
-      }
-      return dtos;
-   }
-   
-   /**
-    * Gets a procedure by UUID.
-    *
-    * GET /rest/v1/emrapi/procedure/{uuid}
-    *
-    * @param uuid the procedure UUID
-    * @return the procedure DTO
-    */
-   @RequestMapping(value = "/{uuid}", method = RequestMethod.GET)
-   @ResponseBody
-   public ResponseEntity<ProcedureDTO> getProcedure(@PathVariable("uuid") String uuid) {
-      Procedure procedure = procedureService.getProcedureByUuid(uuid);
-      if (procedure == null) {
-         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-      }
-      return new ResponseEntity<>(toDTO(procedure), HttpStatus.OK);
-   }
-   
-   /**
-    * Creates a historical procedure.
-    * Historical procedures require an originalDateText value.
-    *
-    * POST /rest/v1/emrapi/procedure/historical
-    *
-    * @param dto the procedure data
-    * @return the created procedure DTO
-    */
-   @RequestMapping(value = "/historical", method = RequestMethod.POST)
-   @ResponseBody
-   public ResponseEntity<ProcedureDTO> createHistoricalProcedure(@RequestBody ProcedureDTO dto) {
-      // TODO: Validate required fields for historical procedure
-      Procedure procedure = fromDTO(dto);
-      procedure = procedureService.saveProcedure(procedure);
-      return new ResponseEntity<>(toDTO(procedure), HttpStatus.CREATED);
-   }
-   
-   /**
-    * Creates a current (non -historical) procedure.
-    * Any originalDateText value is ignored.
-    *
-    * POST /rest/v1/emrapi/procedure/current
-    *
-    * @param dto the procedure data
-    * @return the created procedure DTO
-    */
-   @RequestMapping(value = "/current", method = RequestMethod.POST)
-   @ResponseBody
-   public ResponseEntity<ProcedureDTO> createCurrentProcedure(@RequestBody ProcedureDTO dto) {
-      // TODO: Validate required fields for current procedure
-      Procedure procedure = fromDTO(dto);
-      procedure = procedureService.saveProcedure(procedure);
-      return new ResponseEntity<>(toDTO(procedure), HttpStatus.CREATED);
-   }
-   
-   /**
-    * Converts a Procedure entity to a ProcedureDTO.
-    */
-   private ProcedureDTO toDTO(Procedure procedure) {
-      ProcedureDTO dto = new ProcedureDTO();
-      dto.setUuid(procedure.getUuid());
-      
-      if (procedure.getPatient() != null) {
-         dto.setPatientUuid(procedure.getPatient().getUuid());
-      }
-      if (procedure.getEncounter() != null) {
-         dto.setEncounterUuid(procedure.getEncounter().getUuid());
-      }
-      if (procedure.getProcedureCoded() != null) {
-         dto.setCodedProcedureUuid(procedure.getProcedureCoded().getUuid());
-      }
-      dto.setFreeTextProcedure(procedure.getProcedureNonCoded());
-      
-      if (procedure.getBodySite() != null) {
-         dto.setBodySiteUuid(procedure.getBodySite().getUuid());
-      }
-      
-      dto.setStartDateTime(procedure.getStartDateTime());
-      dto.setEstimatedStartDate(procedure.getEstimatedStartDate());
-      dto.setEndDateTime(procedure.getEndDateTime());
-      dto.setDuration(procedure.getDuration());
-      
-      if (procedure.getDurationUnit() != null) {
-         dto.setDurationUnitUuid(procedure.getDurationUnit().getUuid());
-      }
-      if (procedure.getStatus() != null) {
-         dto.setStatusUuid(procedure.getStatus().getUuid());
-      }
-      if (procedure.getOutcomeCoded() != null) {
-         dto.setCodedOutcomeUuid(procedure.getOutcomeCoded().getUuid());
-      }
-      dto.setFreeTextOutcome(procedure.getOutcomeNonCoded());
-      
-      dto.setNotes(procedure.getNotes());
-      dto.setFormNamespace(procedure.getFormNamespace());
-      dto.setFormFieldPath(procedure.getFormFieldPath());
-      dto.setDateCreated(procedure.getDateCreated());
-      dto.setVoided(procedure.getVoided());
-      
-      return dto;
-   }
-   
-   /**
-    * Creates a new Procedure entity from a ProcedureDTO.
-    */
-   private Procedure fromDTO(ProcedureDTO dto) {
-      Procedure procedure = new Procedure();
-      updateFromDTO(procedure, dto);
-      return procedure;
-   }
-   
-   /**
-    * Updates an existing Procedure entity from a ProcedureDTO.
-    */
-   private void updateFromDTO(Procedure procedure, ProcedureDTO dto) {
-      if (StringUtils.isNotBlank(dto.getPatientUuid())) {
-         Patient patient = patientService.getPatientByUuid(dto.getPatientUuid());
-         if (patient == null) {
-            throw new IllegalArgumentException("Patient not found with UUID: " + dto.getPatientUuid());
-         }
-         procedure.setPatient(patient);
-      }
-      
-      if (StringUtils.isNotBlank(dto.getEncounterUuid())) {
-         Encounter encounter = encounterService.getEncounterByUuid(dto.getEncounterUuid());
-         if (encounter == null) {
-            throw new IllegalArgumentException("Encounter not found with UUID: " + dto.getEncounterUuid());
-         }
-         procedure.setEncounter(encounter);
-      }
-      
-      if (StringUtils.isNotBlank(dto.getCodedProcedureUuid())) {
-         Concept procedureCoded = conceptService.getConceptByUuid(dto.getCodedProcedureUuid());
-         if (procedureCoded == null) {
-            throw new IllegalArgumentException("Procedure concept not found with UUID: " + dto.getCodedProcedureUuid());
-         }
-         procedure.setProcedureCoded(procedureCoded);
-      }
-      procedure.setProcedureNonCoded(dto.getFreeTextProcedure());
-      
-      if (StringUtils.isNotBlank(dto.getBodySiteUuid())) {
-         Concept bodySite = conceptService.getConceptByUuid(dto.getBodySiteUuid());
-         if (bodySite == null) {
-            throw new IllegalArgumentException("Body site concept not found with UUID: " + dto.getBodySiteUuid());
-         }
-         procedure.setBodySite(bodySite);
-      }
-      
-      procedure.setStartDateTime(dto.getStartDateTime());
-      procedure.setEstimatedStartDate(dto.getEstimatedStartDate());
-      procedure.setEndDateTime(dto.getEndDateTime());
-      procedure.setDuration(dto.getDuration());
-      
-      if (StringUtils.isNotBlank(dto.getDurationUnitUuid())) {
-         Concept durationUnitConcept = conceptService.getConceptByUuid(dto.getDurationUnitUuid());
-         if (durationUnitConcept == null) {
-            throw new IllegalArgumentException("Status concept not found with UUID: " + dto.getDurationUnitUuid());
-         }
-         procedure.setDurationUnit(durationUnitConcept);
-      } else {
-         procedure.setDurationUnit(null);
-      }
-      
-      if (StringUtils.isNotBlank(dto.getStatusUuid())) {
-         Concept statusConcept = conceptService.getConceptByUuid(dto.getStatusUuid());
-         if (statusConcept == null) {
-            throw new IllegalArgumentException("Status concept not found with UUID: " + dto.getStatusUuid());
-         }
-         procedure.setStatus(statusConcept);
-      }
-      
-      if (StringUtils.isNotBlank(dto.getCodedOutcomeUuid())) {
-         Concept outcomeCoded = conceptService.getConceptByUuid(dto.getCodedOutcomeUuid());
-         if (outcomeCoded == null) {
-            throw new IllegalArgumentException("Outcome concept not found with UUID: " + dto.getCodedOutcomeUuid());
-         }
-         procedure.setOutcomeCoded(outcomeCoded);
-      }
-      procedure.setOutcomeNonCoded(dto.getFreeTextOutcome());
-      
-      procedure.setNotes(dto.getNotes());
-      
-      if (StringUtils.isNotBlank(dto.getFormNamespace()) && StringUtils.isNotBlank(dto.getFormFieldPath())) {
-         procedure.setFormField(dto.getFormNamespace(), dto.getFormFieldPath());
-      }
-   }
-}
